requirements.txt
Pillow==9.2.0
requests==2.28.1
torch==1.12.0
torchaudio==0.12.0
torchvision==0.13.0
typing_extensions==4.3.0
urllib3==1.26.9
